services/horizon: Fix ReapLookupTables query (#4525)
This commit fixes the query added in ee063a7. The previous query was using
`limit ... offset ...` which becomes slow for larger `offset` values. This is
happening because (from docs[1]):

> The rows skipped by an OFFSET clause still have to be computed inside the
> server; therefore a large OFFSET might be inefficient.

I rewrote the query to use `id` field in tables in `where id > ...`. Before each
run, the new `id` offset is fetched from the table that is later returned to be
used in the next cycle.

I also realized that rewriting `count(*)` in subqueries to `1` with `limit 1`
further improves the query performance. This is because instead of counting all
rows in parent tables it just check if there are _any_ rows in tables (what we
need to determine if the row is orphaned or not). This allows running the reaper
also for `history_accounts` and `history_assets` tables.

Finally this commit adds logging and enable reaper only when
`--history-retention-count` is set.

[1] https://www.postgresql.org/docs/current/queries-limit.html