ingest/ledgerbackend: Skip removing storage dir when Captive-Core in front of requested ledger (#4605)
This commit improves `stellarCoreRunner` code to not remove storage dir if LCL
of Captive-Core is greater than or equal requested `from` ledger.

In e3d3abc we added a code that skips removing storage dir but it was not
removed if and only if `from` argument matched the LCL of Stellar-Core. The
problem is that in clusters with more than one ingesting instance it's possible
that other ingesting node would ingest one or more ledgers while Captive-Core
instance is being restarted. In such case we should not remove storage dir but
just skip the ledgers we don't need.