Remove (unused) stellar/horizon docker image push (#4670)
When migrating the repository from CircleCI to GitHub we were planning
on enabling pushing docker image `stellar/horizon:<commit-sha>` from the
master branch on every merged commit.

However, we never enabled it (note the commented `# push: true` line)

If we haven't used this in 8 months and we haven't missed it, it's
clear we don't need it. Also, `stellar/horizon` is confusing due to
the name similarity with `stellar/stellar-horizon`, see #4669).