ingest/ledgerbackend: Fix bug in PrepareRange (#4772)
There appears to be a bug in PrepareRange where, if you try to use an unbounded range which is ahead of the latest checkpoint, stellar core will try to fetch ledgers from the history archive which don't exist yet. I think this is a change in behavior from stellar core. I assume based on the code comments stellar-core used to wait until the next history archive would be published before attempting to start from a ledger ahead of the latest history archive snapshot.

The bug is fixed by starting core at the latest checkpoint ledger and then seeking ahead to the desired start ledger once the stream is established.