Update location of proto files from protos/ to protos/ingest to be consistent with where the generated code should appear
This also helps simply the Makefile target
Right now, yes, the protos that we need to work on need to be put in the
/ingest path.
Tomorrow, we may need to have generated go files in some other path for some other source code.
At that time, redoing protos might be hard.